Fix the column layout by not using the column layout
The column layout with flexboxes stopped working at some point. I'm not entirely sure why. However, I find the site a little more readable on both desktop and mobile without it, and the content flows fine. Now if only I could figure out how to remove the center-justify from the theme. ;)

With that gone, it appears that most of the scss is useless, so I'm removing it.